# api/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteNetworksBody": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "plz": {
            "title": "Plz",
            "type": "integer"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "plz",
          "version_id"
        ],
        "title": "DeleteNetworksBody",
        "type": "object"
      },
      "DeleteVersionsBody": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "version_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Version Ids",
            "type": "array"
          }
        },
        "required": [
          "version_ids"
        ],
        "title": "DeleteVersionsBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OsmImportBody": {
        "properties": {
          "relation_id": {
            "minimum": 1.0,
            "title": "Relation Id",
            "type": "integer"
          }
        },
        "required": [
          "relation_id"
        ],
        "title": "OsmImportBody",
        "type": "object"
      },
      "PlzBody": {
        "properties": {
          "plz": {
            "title": "Plz",
            "type": "integer"
          }
        },
        "required": [
          "plz"
        ],
        "title": "PlzBody",
        "type": "object"
      },
...
      "SetupBody": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          }
        },
        "title": "SetupBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API of the pylovo workflow (database, regions, data, generation, results) for the GridPlanner UI.",
    "title": "pylovo-api",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/analyze": {
      "post": {
        "description": "``pylovo-analyze --plz <plz> --all`` for the config ``VERSION_ID``.",
        "operationId": "analyze_api_jobs_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlzBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analyze Api Jobs Analyze Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/delete-networks": {
      "post": {
        "description": "``pylovo-delete networks --plz <plz> --version <id>``",
        "operationId": "delete_networks_api_jobs_delete_networks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteNetworksBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Networks Api Jobs Delete Networks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Networks",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/delete-versions": {
      "post": {
        "description": "``pylovo-delete --version <ids>``: removes all results of these versions.",
        "operationId": "delete_versions_api_jobs_delete_versions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteVersionsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Versions Api Jobs Delete Versions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Versions",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/export": {
      "post": {
        "description": "``pylovo-export`` into a private directory; the CSVs are offered for download afterwards.",
        "operationId": "export_api_jobs_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Api Jobs Export Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/generate": {
      "post": {
        "description": "``pylovo-generate --plz ... | --ags ... --parallel|--no-parallel``, gated by the region gate.\n\nPLZ without input data are refused (409) unless ``skip_blocked`` drops them or\n``include_blocked`` keeps them; PLZ whose check has not finished need ``allow_unverified``.\n``--ags`` is passed on only when every PLZ of the municipality passes the gate.",
        "operationId": "generate_api_jobs_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Api Jobs Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/generate/check": {
      "get": {
        "description": "What ``pylovo-generate`` will do for these regions with the current config.",
        "operationId": "generate_check_api_jobs_generate_check_get",
        "parameters": [
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "ags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ags"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Check Api Jobs Generate Check Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Check",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/import-dso": {
      "post": {
        "description": "``pylovo-import transformers-dso-csv <file> [--source S] [--replace-source]``",
        "operationId": "import_dso_api_jobs_import_dso_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DsoImportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Dso Api Jobs Import Dso Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Dso",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/import-osm": {
      "post": {
        "description": "``pylovo-import transformers-osm --relation-id <id>`` (queries the Overpass API).",
        "operationId": "import_osm_api_jobs_import_osm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OsmImportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Osm Api Jobs Import Osm Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Osm",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/setup": {
      "post": {
        "description": "``pylovo-setup --yes``: drops and recreates the ``pylovo`` schema. Type the database name.",
        "operationId": "setup_api_jobs_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Setup Api Jobs Setup Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Setup",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "description": "One job (optionally with the last ``tail`` log lines).",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tail",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 5000,
              "minimum": 0,
              "title": "Tail",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Api Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "description": "Stop a running job (SIGINT to its process group, then SIGTERM/SIGKILL).",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Job Api Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "description": "Server-Sent Events: ``log`` events with new lines, ``status`` updates and a final ``end``.",
        "operationId": "job_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/files": {
      "get": {
        "description": "Files written by an export job.",
        "operationId": "job_files_api_jobs__job_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Job Files Api Jobs  Job Id  Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Files",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}